+++ changelog.md
@@ -1,6 +1,8 @@
 
 # Version 0.1.2
 - added some infrastructure for working with EXTENDED_ARG opcodes 
+- rewritten the re-assembler of the instructions into the real bytecode 
+  - EXTENDED_ARG opcodes are now skipped during disassembly and added back during assembling
 
 # Version 0.1.0 / 0.1.1